fix data type register bug; test=develop

+++ paddle/pten/kernels/gpu/transpose_grad_kernel.cu
@@ -27,6 +27,6 @@ PT_REGISTER_KERNEL(transpose_grad,
                    double,
                    int32_t,
                    int64_t,
-                   pten::dtype::bfloat16,
+                   pten::dtype::float16,
                    paddle::platform::complex<float>,
                    paddle::platform::complex<double>) {}
